SPA - migrate maintenance page
